requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
Django==3.0.6
django-allauth==0.42.0
django-ckeditor==5.9.0
django-cors-headers==3.3.0
django-filter==2.2.0
django-js-asset==1.2.2
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
drf-yasg==1.17.1
get==2019.4.13
idna==2.9
inflection==0.4.0
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
oauthlib==3.1.0
packaging==20.4
Pillow==7.1.2
pip==20.1.1
post==2019.4.13
psycopg2-binary==2.8.5
public==2019.4.13
PyJWT==1.7.1
pyparsing==2.4.7
python3-openid==3.1.0
pytz==2020.1
query-string==2019.4.13
requests==2.23.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
setuptools==46.4.0
six==1.15.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
